epicEvents/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.4
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
flake8==6.1.0
flake8-html==0.4.3
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-mock==3.12.0
pytz==2023.3.post1
requests==2.31.0
rich==13.7.0
sentry-sdk==1.39.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
